fetch/metadata/generated/redirect-cross-site-cross-site.tentative.https.image.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <body>
  <script>
  'use strict';
  let nonce = '{{uuid()}}';
  promise_test(() => {
    const path = '/fetch/metadata/resources/record-header.py?file=' + nonce;

    return new Promise((resolve) => {
      const image = document.createElement('img');
      image.src = 'https://{{hosts[alt][www]}}:{{ports[https][0]}}/xhr/resources/redirect.py?location=https://{{hosts[alt][www]}}:{{ports[https][0]}}' + path;
      image.onload = image.onerror = resolve;
    })
    .then(() => fetch(path + '&retrieve'))
    .then((response) => response.text())
    .then((text) => JSON.parse(text))
    .then((actual) => {
let expected = {
  site: "cross-site",
  user: "",
  mode: "no-cors",
  dest: "image"
};
assert_header_equals(actual, expected, "Cross-Site -> Cross-Site redirect");

    });
  }, 'Cross-Site -> Cross-Site redirect');
  </script>
  </body>
</html>
